refactor user service

+++ User/service/views.py
@@ -4,21 +4,26 @@ from rest_framework.response import Response
 from rest_framework import status
 
 from .models import UserModel
-from .serializers import UserSerializer
-from  .publisher import publish_message
+from .serializers import UserSerializer, UserListSerializer
+from .publisher import publish_message
 
 
 @api_view(['GET'])
-def getAllUser(request):
+def get_all_user(request):
     users = UserModel.objects.all()
-    serializer = UserSerializer(users, many=True)
+    serializer = UserListSerializer(users, many=True)
 
     return Response(serializer.data, status=status.HTTP_200_OK)
+
 
 @api_view(['POST'])
 def login(request):
     try:
-        user = UserModel.objects.get(email=request.data['email'])
+        filtered_user = UserModel.objects.filter(email=request.data['email'])
+        if not filtered_user:
+            return Response("no such user", status=status.HTTP_403_FORBIDDEN)
+        user = filtered_user[0]
+
         if user.password == request.data['password']:
             return Response(user.userId, status=status.HTTP_200_OK)
         else:
@@ -38,29 +43,29 @@ def admin_login(request):
 
 
 @api_view(['GET'])
-def getUser(request, user_id):
-    user = UserModel.objects.filter(userId = user_id)
+def get_user(request, user_id):
+    user = UserModel.objects.filter(userId=user_id)
     serializer = UserSerializer(user, many=True)
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 @api_view(['POST'])
-def createUser(request):
-    users = UserModel.objects.filter(email = request.data['email'])
+def create_user(request):
+    users = UserModel.objects.filter(email=request.data['email'])
 
     try:
         if len(users):
             raise Exception("Email has been used")
 
-        serializer = UserSerializer(data = request.data)
+        serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             serialized_data = serializer.data
             user = UserModel(
-                userId = serialized_data['userId'],
-                name = serialized_data['name'],
-                email = serialized_data['email'],
-                password = serialized_data['password']
+                userId=serialized_data['userId'],
+                name=serialized_data['name'],
+                email=serialized_data['email'],
+                password=serialized_data['password']
             )
             publish_message("POST", user)
 
@@ -72,24 +77,24 @@ def createUser(request):
 
 
 @api_view(['PUT'])
-def updateUser(request, user_id):
+def update_user(request, user_id):
     try:
-        user = UserModel.objects.filter(email = request.data['email'])
+        user = UserModel.objects.filter(email=request.data['email'])
         if len(user) > 0:
             for data in user.values():
                 if data['userId'] != user_id:
                     raise Exception("Email already exists")
-                
-        user = UserModel.objects.get(userId = user_id)
-        serializer = UserSerializer(user, data = request.data)
+
+        user = UserModel.objects.get(userId=user_id)
+        serializer = UserSerializer(user, data=request.data)
 
         if serializer.is_valid():
             serializer.save()
             user = UserModel(
-                userId = serializer.data['userId'],
-                name = serializer.data['name'],
-                email = serializer.data['email'],
-                password = serializer.data['password']
+                userId=serializer.data['userId'],
+                name=serializer.data['name'],
+                email=serializer.data['email'],
+                password=serializer.data['password']
             )
             publish_message("PUT", user)
         return Response(serializer.data, status=status.HTTP_200_OK)
@@ -98,9 +103,13 @@ def updateUser(request, user_id):
 
 
 @api_view(['DELETE'])
-def deleteUser(request, user_id):
+def delete_user(request, user_id):
     try:
-        user = UserModel.objects.get(userId = user_id)
+        filtered_user = UserModel.objects.filter(userId=user_id)
+        if not filtered_user:
+            return Response("no such user", status=status.HTTP_403_FORBIDDEN)
+        user = filtered_user[0]
+
         publish_message("DELETE", user)
         user.delete()
         return Response("", status=status.HTTP_200_OK)
